Folder delete시 "분류되지 않음" 폴더로 문제집 이동 구현 기타 에러 수정

+++ s-with-me-back/src/main/java/com/swithme/web/dto/MyBookResponseDto.java
@@ -11,10 +11,10 @@ public class MyBookResponseDto {
     private int myBookId;
     private int folderId;
     private int bookId;
-    private short lastProblemId;
+    private int lastProblemId;
 
     @Builder
-    public MyBookResponseDto(int myBookId, int folderId, int bookId, short lastProblemId){
+    public MyBookResponseDto(int myBookId, int folderId, int bookId, int lastProblemId){
         this.myBookId = myBookId;
         this.folderId = folderId;
         this.bookId = bookId;
